+++ Database/Migrations/2026_03_15_000001_create_scheduled_conversations_table.php
@@ -15,7 +15,6 @@
  *
  * @package Modules\ScheduledConversations
  * @author  Raimundo Alba
- * @version 1.6.0
  */
 
 use Illuminate\Support\Facades\Schema;