+++ Entities/ScheduledConversationLog.php
@@ -18,7 +18,6 @@
  *
  * @package Modules\ScheduledConversations
  * @author  Raimundo Alba
- * @version 1.6.0
  */
 
 namespace Modules\ScheduledConversations\Entities;